Split forms, services and views into different files
