+++ conventions/views/convention_finalisation.py
@@ -1,118 +1,21 @@
-from django import forms
-from django.http import HttpRequest, HttpResponseRedirect
+from django.conf import settings
+from django.http import HttpResponseRedirect
 from django.urls import reverse
 from django.views.generic.base import TemplateView
+from waffle.mixins import WaffleFlagMixin
 
-from conventions.models.convention import Convention
 from conventions.services import utils
+from conventions.services.finalisation import (
+    FinalisationCerfaService,
+    FinalisationNumeroService,
+    FinalisationServiceBase,
+    FinalisationValidationService,
+)
 from core.stepper import Stepper
 
 
-class FinalisationFormBase(forms.Form):
-    pass
-
-
-class FinalisationNumeroForm(FinalisationFormBase):
-    uuid = forms.UUIDField(
-        required=False,
-        label="Finalisation numéro",
-    )
-    numero = forms.CharField(
-        label="Numéro de convention",
-        help_text="Cet identifiant proposé est unique et standardisé à l'échelle nationale."
-        '<a href="https://siap-logement.atlassian.net/wiki/x/f4Bu">En savoir plus</a>',
-        max_length=255,
-        min_length=1,
-        required=True,
-        error_messages={
-            "required": "Le numéro de la convention est obligatoire",
-            "min_length": "Le numéro de la convention est obligatoire",
-            "max_length": "Le numéro de la convention ne doit pas excéder 255 caractères",
-        },
-    )
-
-
-class FinalisationCerfaForm(FinalisationFormBase):
-    uuid = forms.UUIDField(
-        required=False,
-        label="Finalisation cerfa",
-    )
-    fichier_override_cerfa = forms.CharField(required=False, label="Cerfa personalisé")
-    fichier_override_cerfa_files = forms.CharField(
-        required=False,
-        help_text="Les fichiers de type docx sont acceptés dans la limite de 100 Mo",
-    )
-
-
-class FinalisationServiceBase:
-    form: FinalisationFormBase
-    convention: Convention
-
-    def __init__(self, convention_uuid: str, request: HttpRequest) -> None:
-        self.convention = Convention.objects.get(uuid=convention_uuid)
-        if request.method == "POST":
-            self.form = FinalisationNumeroForm(request.POST, request.FILES)
-        else:
-            self.form = FinalisationNumeroForm(
-                initial={"numero": self.convention.numero}
-            )
-
-
-class FinalisationNumeroService(FinalisationServiceBase):
-    form: FinalisationNumeroForm
-
-    def save(self) -> str:
-        if self.form.is_valid():
-            self.convention.numero = self.form.cleaned_data["numero"]
-            self.convention.save()
-            return utils.ReturnStatus.SUCCESS
-        return utils.ReturnStatus.ERROR
-
-
-class FinalisationCerfaService(FinalisationServiceBase):
-    form: FinalisationCerfaForm
-
-    def __init__(self, convention_uuid: str, request: HttpRequest) -> None:
-        self.convention = Convention.objects.get(uuid=convention_uuid)
-
-        if request.method == "POST":
-            self.form = FinalisationCerfaForm(
-                {
-                    "uuid": self.convention.uuid,
-                    **utils.init_text_and_files_from_field(
-                        request,
-                        self.convention,
-                        "fichier_override_cerfa",
-                    ),
-                }
-            )
-        else:
-            self.form = FinalisationCerfaForm(
-                initial={
-                    "uuid": self.convention.uuid,
-                    **utils.get_text_and_files_from_field(
-                        "fichier_override_cerfa",
-                        self.convention.fichier_override_cerfa,
-                    ),
-                }
-            )
-
-    def save(self) -> str:
-        if self.form.is_valid():
-            self.convention.fichier_override_cerfa = utils.set_files_and_text_field(
-                self.form.cleaned_data["fichier_override_cerfa_files"],
-                self.form.cleaned_data["fichier_override_cerfa"],
-            )
-            self.convention.save()
-            return utils.ReturnStatus.SUCCESS
-        return utils.ReturnStatus.ERROR
-
-
-class FinalisationValidationService(FinalisationServiceBase):
-    pass
-
-
-class FinalisationBase(TemplateView):
+class FinalisationBase(WaffleFlagMixin, TemplateView):
+    waffle_flag = settings.FLAG_OVERRIDE_CERFA
     service_class: FinalisationServiceBase
 
     def __init__(self, *args, **kwargs) -> None:
